requirements.txt
fa2==0.3.5
jupyter==1.0.0
matplotlib==3.2.2
matplotlib_venn==0.11.5
minet==0.34.0
networkx==2.5
nltk==3.5
numpy==1.19.0
pandas==1.0.5
wordcloud==1.8.1
python-dotenv==0.15.0
requests==2.25.0